backend/pyproject.toml
[tool.poetry]
name = "fastapi-seed"
version = "0.1.0"
description = "FastAPI Boilerplate"
authors = ["loctvl842 <loclepnvx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
toml = "^0.10.2"
loguru = "^0.7.2"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.4"
supervisor = "^4.2.5"
httpx = "^0.27.2"
tenacity = "^9.0.0"
dramatiq = "^1.17.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
poetry-plugin-export = "^1.8.0"
ujson = "^5.10.0"
psycopg2 = "^2.9.10"
fastapi-mail = "^1.4.2"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
